Add scalable app structure and divide webpack config per environment
