+++ .babelrc
@@ -4,5 +4,5 @@
     "react",
     "stage-0"
   ],
-  "plugins": ["transform-runtime", "react-hot-loader/babel"]
+  "plugins": ["transform-runtime", "transform-decorators-legacy", "react-hot-loader/babel"]
 }
