+++ dev-server.js
@@ -1,3 +1,5 @@
+// Creates a hot reloading development environment
+
 const path = require('path');
 const express = require('express');
 const webpack = require('webpack');
@@ -32,7 +34,7 @@ app.use(webpackDevMiddleware(compiler, {
 app.use(webpackHotMiddleware(compiler));
 
 app.get('*', (req, res) => {
-  res.sendFile(path.join(__dirname, './src/index.html'));
+  res.sendFile(path.join(__dirname, './src/client/assets/index.html'));
 });
 
 app.listen(port, host, (err) => {
