+++ package.json
@@ -38,6 +38,7 @@
     "babel-core": "6.13.2",
     "babel-eslint": "6.1.2",
     "babel-loader": "6.2.4",
+    "babel-plugin-transform-decorators-legacy": "1.3.4",
     "babel-plugin-transform-runtime": "6.12.0",
     "babel-preset-es2015-native-modules": "6.9.3",
     "babel-preset-react": "6.5.0",
@@ -90,6 +91,7 @@
     "react-router-redux": "4.0.5",
     "redux": "3.5.2",
     "redux-logger": "2.6.1",
-    "redux-promise": "0.5.3"
+    "redux-promise": "0.5.3",
+    "reselect": "2.5.3"
   }
 }
