+++ src/client/assets/images/favicon.ico
