+++ src/client/assets/javascripts/app/App.js
@@ -1,0 +1,13 @@
+import React, { PropTypes } from 'react';
+
+const App = ({ children }) => (
+  <div className="page-container">
+    {children}
+  </div>
+);
+
+App.propTypes = {
+  children: PropTypes.element.isRequired
+};
+
+export default App;
