+++ src/client/assets/javascripts/app/DevTools.js
