+++ src/client/assets/javascripts/app/store/configureStore.js
