+++ src/client/assets/javascripts/features/friends/components/AddFriendInput/index.js
@@ -1,0 +1,1 @@
+export { default } from './AddFriendInput';
