+++ src/client/assets/javascripts/features/friends/components/FriendList/FriendList.scss
