+++ src/client/assets/javascripts/features/friends/components/FriendList/index.js
@@ -1,0 +1,1 @@
+export { default } from './FriendList';
