+++ src/client/assets/javascripts/features/friends/components/FriendListApp.scss
