+++ src/client/assets/javascripts/features/friends/components/FriendListItem/FriendListItem.scss
