+++ src/client/assets/javascripts/features/friends/components/FriendListItem/fonts/opensans-regular-webfont.woff
