+++ src/client/assets/javascripts/features/friends/components/FriendListItem/images/icon-user.png
