+++ src/client/assets/javascripts/features/friends/components/FriendListItem/index.js
@@ -1,0 +1,1 @@
+export { default } from './FriendListItem';
