+++ src/client/scripts/development.js
@@ -1,0 +1,7 @@
+// Load CSS via Webpack to be able to require Bootstrap, Font Awesome, etc. from npm
+// require('./development.scss');
+
+require('styles/styles.scss');
+
+// JavaScript main file
+require('javascripts/app');
