+++ src/client/scripts/production.js
@@ -1,0 +1,8 @@
+// Expose Raven
+// window.Raven = require('raven-js');
+
+// Load CSS via Webpack to be able to require Bootstrap, Font Awesome, etc. from npm
+// require('./production.scss');
+
+// JavaScript main file
+// require('../assets/javascripts/app');
