--- src/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Redux minimal boilerplate</title>
-  <link rel="shortcut icon" href="/images/favicon.ico">
-  <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/font-awesome/4.3.0/css/font-awesome.min.css">
-</head>
-<body>
-
-<div id="app"></div>
-
-<script src="/js/vendor.bundle.js"></script>
-<script src="/js/application.js"></script>
-</body>
-</html>
