--- src/js/components/index.js
@@ -1,2 +1,0 @@
-export { default as AddFriendInput } from './AddFriendInput/AddFriendInput';
-export { default as FriendList } from './FriendList/FriendList';
