--- src/js/routes.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import { Route, IndexRoute, Redirect } from 'react-router';
-
-import App from './components/App';
-import FriendListApp from './containers/FriendListApp/FriendListApp';
-import NotFoundView from './views/NotFoundView';
-
-export default (
-  <Route path="/" component={App}>
-    <IndexRoute component={FriendListApp} />
-    <Route path="404" component={NotFoundView} />
-    <Redirect from="*" to="404" />
-  </Route>
-);
